VeroctaAIApp/requirements.txt
# VeroctaAI Financial Intelligence Platform - Core Dependencies
# Web framework
flask==3.1.1
werkzeug==3.1.3

# Environment variables
python-dotenv==1.0.1

# Data processing
pandas==2.3.1
numpy>=1.26.0

# AI integration
openai==1.98.0

# PDF generation
reportlab==4.4.3
matplotlib==3.10.5

# Image processing
Pillow>=10.0.0

# Production server
gunicorn==23.0.0

# Date utilities
python-dateutil>=2.8.2
